feat!: add loki with multi-tenancy for notifications
This adds bitnami's loki helm chart, two instances of promtail
(one for logs and one for notifications on kafka)
and a self-maintained loki-gateway to allow multiple tenants.

Promtail prevents this from working currently as it's not allowed
to inject a secret with release name in through the volumes.

Ideally also an init container is added to promtail-notifications
as it crashes until kafka is available.
This is currently also prevented by not being able to inject templates in the init container.
Even if the init container would be there though,
there should be data on kafka before promtail will report the pod as ready.
This will need to be solved before the github actions workflows will pass.