feat: allow to pull external (OCI) charts
To pull external charts in (only from an OCI registry),
we need a chart in the `charts` dir,
but the contents should only be a single file `External.yaml`.
Then the `External.yaml` file should contain someting similar to
a `dependencies` entry found in a normal `Chart.yaml` file:

```yaml
name: cu-cp-chart
alias: cu-cp
repository: oci://ghcr.io/accelleran
version: 12.0.0
```

`alias` is an optional field in case we want to use a different name as the original chart.

The repository name is also a proposal for how we organize ourselves with OCI charts.
We could also use docker hub, though note that an explicit registry is required for helm as it has no default fallback.
Also docker doesn't support using a `/` in the repository (contrary to ghcr),
so we can't for example use a repo prefix like `accelleran/helm-charts`
that would look like `oci://ghcr.io/accelleran/helm-charts/cu-cp` in the end result.
This means that the chart name needs to be different than the docker image name,
hence the `-chart` suffix.
I would still adapt to these limitations when using ghcr
to keep the option open for both registries in a consistent way.

Note also that for renovate we should use a specific version,
so we can't use a version range like we otherwise could with helm.
Also the order of all the keys is important as a regex manager is used within renovate.

I've also added a package rule for external charts to let renovate immediately automerge them.
Here I make the assumption that the external chart release process has tested the chart properly.
To further support this we could extract some (composite) actions
from the current workflow for easy use in this repo and externally.

Currently I've also made up something to login to ghcr.
GitHup Apps are not supported at the moment to login to ghcr,
so we'll need to see how we access the external charts if using ghcr.