fix: fix accessing drax api with access token (#1202)
* fix: fix adding role to client scope

* fix: fix client scope log messages

* fix: add oauth2-proxy audience to external client access tokens

* fix: set default assigned type for oauth2-proxy client scope

* fix: ignore client scope mapper updates

There seems to be a bug with a NullPointerException in keycloak,
although we maybe just set something wrong.
We get into this state also through the UI,
so for now we just ignore updates failing.
Once we actually would want to perform an update
on a client scope mapper in a future version,
we'll need to properly look into this.